+++ coordinator/coordinator.go
@@ -242,6 +242,9 @@ func (self *Coordinator) getShardsAndProcessor(querySpec *parser.QuerySpec, writ
 			shardIds[i] = s.Id()
 		}
 		writer, err = engine.NewQueryEngine(writer, q, shardIds)
+		if err != nil {
+			log.Info("Coordinator processor chain: %s", engine.ProcessorChain(writer))
+		}
 		return shards, writer, err
 	}