+++ integration/data_test.go
@@ -550,6 +550,41 @@ func (self *DataTestSuite) TestWhereAndArithmetic(c *C) {
 }
 
 // issue #524
+func (self *DataTestSuite) TestJoinRegex(c *C) {
+	t1 := time.Now().Truncate(time.Hour).Add(-4 * time.Hour)
+	t2 := t1.Add(time.Hour)
+	data := fmt.Sprintf(`[
+{
+  "name":"foo1",
+  "columns":["time", "val"],
+  "points":[[%d, 1],[%d, 2]]
+},
+{
+  "name":"foo2",
+  "columns":["time", "val"],
+  "points":[[%d, 3],[%d, 4]]
+
+},
+{
+  "name":"foo3",
+  "columns":["time", "val"],
+  "points":[[%d, 5],[%d, 6]]
+
+}]`, t1.Unix(), t2.Unix(), t1.Unix(), t2.Unix(), t1.Unix(), t2.Unix())
+	self.client.WriteJsonData(data, c, "s")
+	serieses := self.client.RunQuery("select * from join(/foo\\d+/)", c, "m")
+	c.Assert(serieses, HasLen, 1)
+	maps := ToMap(serieses[0])
+	c.Assert(maps, HasLen, 2)
+	c.Assert(maps[0]["foo1.val"], Equals, 2.0)
+	c.Assert(maps[0]["foo2.val"], Equals, 4.0)
+	c.Assert(maps[0]["foo3.val"], Equals, 6.0)
+	c.Assert(maps[1]["foo1.val"], Equals, 1.0)
+	c.Assert(maps[1]["foo2.val"], Equals, 3.0)
+	c.Assert(maps[1]["foo3.val"], Equals, 5.0)
+}
+
+// issue #524
 func (self *DataTestSuite) TestJoinAndArithmetic(c *C) {
 	t1 := time.Now().Truncate(time.Hour).Add(-4 * time.Hour)
 	t2 := t1.Add(time.Hour)